refactor(AIChat): 重构manager结构
