+++ modules/self_contained/ai_chat/providers/__init__.py
@@ -1,0 +1,19 @@
+"""
+AI提供商模块
+"""
+
+# 这里可以导入各种AI提供商实现
+# 例如：from .openai_provider import OpenAIProvider
+
+# 导出所有提供商类，便于外部直接从模块导入
+# __all__ = ["OpenAIProvider", "AnthropicProvider", ...]
+
+
+from .deepseek import DeepSeekProvider
+from .openai import OpenAIProvider
+
+__all__ = [
+    "OpenAIProvider",
+    "DeepSeekProvider",
+    # 添加其他提供商类
+]
